Validations added.

+++ Dissertation/main/Queries.java
@@ -2,6 +2,8 @@ package main;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.Random;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import javax.swing.JLabel;
 import javax.swing.JMenu;
 import javax.swing.JOptionPane;
@@ -272,32 +274,52 @@ public class Queries {
 	//Remove course selected by student
 	//String s passes the course to be removed
 	//int d passes the studentID
-	public void removeChoice(String s, int d){
+	public String removeChoice(String s, int d){
+		String get = "";
+		ResultSet rs;
 		try{
+			String query = "SELECT * FROM COURSERESULT WHERE COURSE = '"+s+"' AND STUDENTID = '"+d+"' AND RESULT IS NOT NULL";
 			st = conn.createStatement();
-			String sql = "DELETE FROM COURSERESULT WHERE COURSE = '"+s+"' AND STUDENTID = '"+d+"'";
-			st.executeUpdate(sql);
+			rs = st.executeQuery(query);
+
+			if(rs.next()){
+				get+="Exists";
+			}else{
+				String sql = "DELETE FROM COURSERESULT WHERE COURSE = '"+s+"' AND STUDENTID = '"+d+"'";
+				st.executeUpdate(sql);
+			}
+			rs.close();
+			st.close();
 		}catch (Exception e){
 			e.printStackTrace();
 		}
+
+		return get;
 	}
 	//Insert new courses into the system by admin
 	public String insertCourse(String s, int d, int d1, int d2){
 		String check = "";
+		int count = 0;
 		ResultSet rs;
 
 		try{
 			st = conn.createStatement();
-			String query = "SELECT COURSENAME FROM COURSES WHERE COURSENAME = '"+s+"'";
+			String query = "SELECT COURSENAME FROM COURSES";
 			rs = st.executeQuery(query);
-			if(rs.next()){
-				check+="Error";
-
-			}else{
-				String sql = "INSERT INTO COURSES VALUES ('"+s+"','"+d+"','"+d1+"','"+d2+"')";
-				st.executeUpdate(sql);
+			while(rs.next()){
+				String name = rs.getString("COURSENAME");
+				if(s.toLowerCase().equals(name.toLowerCase())){
+					count+=1;
+					check+="Error";
+				}
 
 			}
+			if(count == 0){
+				String sql = "INSERT INTO COURSES VALUES ('"+s+"','"+d+"','"+d1+"','"+d2+"')";
+				st.executeUpdate(sql);
+			}
+
+
 
 			rs.close();
 			st.close();
@@ -322,8 +344,8 @@ public class Queries {
 				if(rs.next()){
 					sb.append("Error");
 				}else{
-				String sql = "INSERT INTO COURSEDEGREE VALUES ('"+course+"', '"+degreeid+"')";
-				st.executeUpdate(sql);
+					String sql = "INSERT INTO COURSEDEGREE VALUES ('"+course+"', '"+degreeid+"')";
+					st.executeUpdate(sql);
 				}
 			}
 			rs.close();
@@ -334,16 +356,27 @@ public class Queries {
 		return sb.toString();
 	}
 	//Remove course from the system
-	public void removeCourse(String s){
+	public String removeCourse(String s){
+		ResultSet rs;
+		String get = "";
 		try{
+			String query = "SELECT * FROM COURSERESULT WHERE COURSE = '"+s+"'";
 			st = conn.createStatement();
-			String sql1 = "DELETE FROM COURSEDEGREE WHERE COURSE_NAME = '"+s+"'";
-			st.executeUpdate(sql1);
-			String sql = "DELETE FROM COURSES WHERE COURSENAME = '"+s+"'";
-			st.executeUpdate(sql);
+			rs = st.executeQuery(query);
+			if(rs.next()){
+				get+="Exists";
+			}else{
+				String sql1 = "DELETE FROM COURSEDEGREE WHERE COURSE_NAME = '"+s+"'";
+				st.executeUpdate(sql1);
+				String sql = "DELETE FROM COURSES WHERE COURSENAME = '"+s+"'";
+				st.executeUpdate(sql);
+			}
+			rs.close();
+			st.close();
 		}catch (Exception e){
 			e.printStackTrace();
 		}
+		return get;
 	}
 	//Display courses available for selection
 	//int s takes the studentID
@@ -399,7 +432,7 @@ public class Queries {
 			String queryx = "SELECT DISTINCT COURSENAME FROM COURSES AS c INNER JOIN COURSEDEGREE AS cd ON c.COURSENAME = cd.COURSE_NAME INNER JOIN "
 					+ "DEGREE AS d ON d.DEGREEID = cd.DEGREE_ID INNER JOIN SCHOOL AS s ON s.SCHOOLNAME = d.SCHOOL_REF INNER JOIN "
 					+ "ADMIN AS ad ON ad.SCHOOLREF = s.SCHOOLNAME INNER JOIN USER AS us ON us.ID = ad.USERID WHERE us.MATRICNO = '"+s+"' ORDER BY COURSENAME ASC";
-			
+
 			st = conn.createStatement();
 			rs = st.executeQuery(queryx);
 
@@ -496,6 +529,9 @@ public class Queries {
 		int totalcred = 0, totalpoints = 0;
 		double average = 0;
 		int count = 0;
+		int p = 0;
+		String score = "";
+		String total = "";
 		int id = getStudent(fname, lname);
 		String names = String.format("%s %s\n", fname, lname);
 		sb.append(names+"\n");
@@ -505,7 +541,7 @@ public class Queries {
 		ResultSet rs;
 
 		try{
-			String query = "SELECT cr.COURSE, cr.RESULT, c.CREDIT FROM COURSERESULT AS cr INNER JOIN COURSES AS c ON c.COURSENAME = cr.COURSE WHERE cr.STUDENTID = '"+id+"'";
+			String query = "SELECT cr.COURSE, cr.RESULT, c.CREDIT FROM COURSERESULT AS cr INNER JOIN COURSES AS c ON c.COURSENAME = cr.COURSE WHERE cr.STUDENTID = '"+id+"' ORDER BY cr.COURSE ASC";
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
 			while(rs.next()){
@@ -514,30 +550,36 @@ public class Queries {
 				String course = rs.getString("COURSE");
 				int res = rs.getInt("RESULT");
 				int credit = rs.getInt("CREDIT"); 
+				int r = res * credit;
+				p+=res;
 				if(res == 0){
-					res = 0;
-				}
-
-
-				String s = String.format(" %-50.50s %-10s %-10s %-10s\n", course, credit, res, (res*credit));
-				sb.append(s);
-				sb.append("\n");
-				if(res != 0){
-					totalpoints+=(res*credit);
+					score = "";
+					total = "";
+				}else{
+					score = getScore(res);
+					total = ""+(r)+"";
 					totalcred+=credit;
 				}
+				String s = String.format(" %-50.50s %-10d %-10s %-10s\n", course, credit, score, total);
+				sb.append(s);
+				sb.append("\n");
+				totalpoints+=(r);
+
 
 			}
 			average+=((double)totalpoints/totalcred);
 
 			if (count == 0){
-				return "Not Available";
+				return "No courses selected.";
 			}
-			String s2 = String.format(" %65s:%5d\n", "Total", totalpoints);
-			sb.append(s2);
-			sb.append("\n");
-			String s1 = String.format("Your overall result is %d/%d : %.3f", totalpoints, totalcred, average);
-			sb.append(s1);
+
+			if(p != 0){
+				String s2 = String.format(" %65s:%5d\n", "Total", totalpoints);
+				sb.append(s2);
+				sb.append("\n");
+				String s1 = String.format("Your overall result is %d/%d : %.3f", totalpoints, totalcred, average);
+				sb.append(s1);
+			}
 
 			rs.close();
 			st.close();
@@ -649,16 +691,14 @@ public class Queries {
 		ResultSet rs;
 		try{
 			int id = getStudent(fname, lname);
-			String query1 = "SELECT cr.COURSE, cr.RESULT, c.CREDIT FROM COURSERESULT AS cr INNER JOIN COURSES AS c ON c.COURSENAME = cr.COURSE INNER JOIN STUDENT AS s ON s.STUDENTID = cr.STUDENTID WHERE s.STUDENTID = '"+id+"'";
+			String query1 = "SELECT cr.RESULT FROM COURSERESULT AS cr INNER JOIN COURSES AS c ON c.COURSENAME = cr.COURSE INNER JOIN STUDENT AS s ON s.STUDENTID = cr.STUDENTID WHERE s.STUDENTID = '"+id+"' AND cr.result IS NULL";
 			st = conn.createStatement();
 			rs = st.executeQuery(query1);
 			while(rs.next()){
 				count+=1;
-				int res = rs.getInt("RESULT");
-				check+=res;
-				
+
 			}
-			if(check == 0 || count == 0){
+			if(count > 0){
 				sb.append("No");
 			}
 			rs.close();
@@ -716,7 +756,7 @@ public class Queries {
 		}
 		return id;
 	}
-	
+
 	public String overallCourse(String s){
 		StringBuilder sb = new StringBuilder("");
 		int totalpoints = 0;
@@ -725,7 +765,7 @@ public class Queries {
 		int numOfStudents = 0;
 		String degree = String.format("%s\n", s);
 		sb.append(degree + "\n");
-		String display = String.format(" %-50.50s %-10s\n", "Name", "Overall Mark");
+		String display = String.format(" %-40.40s %-10s\n", "Name", "Overall Mark");
 		sb.append(display);
 		sb.append("\n");
 		ResultSet rs;
@@ -735,27 +775,30 @@ public class Queries {
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
 			while(rs.next()){
-				
+
 				String fname = rs.getString("FIRSTNAME");
 				String lname = rs.getString("LASTNAME");
 				int res = rs.getInt("RESULT");
-				 
+				count+=res;
 				if(res != 0){
 					String name = fname +" "+ lname;
 					numOfStudents+=1;
-					String s1 = String.format(" %-50.50s %-10d\n", name, res);
+					String mark = getScore(res);
+					String s1 = String.format(" %-40.40s %-10s\n", name, mark);
 					sb.append(s1);
 					sb.append("\n");
 					totalpoints+=res;
-				}else{
-					count+=1;
 				}
 
 			}
-			average+=((double)totalpoints/numOfStudents);
-			
-			String s1 = String.format("The overall result for this course %.3f", average);
-			sb.append(s1);
+
+			if(count !=0){
+				average+=((double)totalpoints/numOfStudents);
+				String s1 = String.format("The overall average for this course is %.3f", average);
+				sb.append(s1);
+			}else{
+				sb.append("No marks available yet.");
+			}
 
 			rs.close();
 			st.close();
@@ -765,39 +808,39 @@ public class Queries {
 		}
 		return sb.toString();
 	}
-	
+
 	public String allStudents(int id){
 		ResultSet rs;
 		StringBuilder sb = new StringBuilder("");
 		String display = String.format(" %-40.40s %-20s %-40.40s %-10s\n", "Name","Matric Number","Degree" ,"Email");
 		sb.append(display);
-		
+
 		try{
 			String query = "SELECT s.FIRSTNAME, s.LASTNAME, d.DEGREENAME, s.EMAIL FROM STUDENT AS s INNER JOIN STUDENT_DEGREE AS sd "
-				+ "ON sd.STUDENT = s.STUDENTID INNER JOIN DEGREE AS d ON d.DEGREEID = sd.DEGREE INNER JOIN ADMIN AS ad ON ad.SCHOOLREF = d.SCHOOL_REF WHERE ad.ADMINID = '"+id+"'";
+					+ "ON sd.STUDENT = s.STUDENTID INNER JOIN DEGREE AS d ON d.DEGREEID = sd.DEGREE INNER JOIN ADMIN AS ad ON ad.SCHOOLREF = d.SCHOOL_REF WHERE ad.ADMINID = '"+id+"'";
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
-			
+
 			while(rs.next()){
 				String fname = rs.getString("FIRSTNAME");
 				String lname = rs.getString("LASTNAME");
 				String degree = rs.getString("DEGREENAME");
 				String email = rs.getString("EMAIL");
-				
+
 				String matric = matric(fname, lname);
 				String name = fname+" "+lname;
-				
+
 				String area = String.format(" %-40.40s %-20s %-40.40s %-10s\n", name, matric, degree, email);
 				sb.append(area+"\n");
 			}
 			rs.close();
 			st.close();
 		}catch(Exception e){
-			
+
 		}
 		return sb.toString();
 	}
-	
+
 	public String overallSchool(int id){
 		Double score = 0.0;
 		StringBuilder sb = new StringBuilder("");
@@ -809,27 +852,27 @@ public class Queries {
 		ResultSet rs;
 		try{
 			String query = "SELECT s.FIRSTNAME, s.LASTNAME, d.DEGREENAME, sd.RESULT FROM STUDENT_DEGREE AS sd INNER JOIN DEGREE AS d "
-				+ "ON d.DEGREEID = sd.DEGREE INNER JOIN ADMIN AS a ON a.SCHOOLREF = d.SCHOOL_REF INNER JOIN STUDENT AS s ON s.STUDENTID = sd.STUDENT WHERE a.ADMINID = '"+id+"' AND sd.RESULT IS NOT NULL";
+					+ "ON d.DEGREEID = sd.DEGREE INNER JOIN ADMIN AS a ON a.SCHOOLREF = d.SCHOOL_REF INNER JOIN STUDENT AS s ON s.STUDENTID = sd.STUDENT WHERE a.ADMINID = '"+id+"' AND sd.RESULT IS NOT NULL";
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
-			
+
 			if(rs.next()){
 				count+=1;
 				String fname = rs.getString("FIRSTNAME");
 				String lname = rs.getString("LASTNAME");
 				String degreename = rs.getString("DEGREENAME");
 				String result = rs.getString("RESULT");
-				
+
 				String name = fname +" "+ lname;
 				Double res = Double.parseDouble(result);
 				String names = String.format(" %-40.40s %-30.30s %-10.3f\n", name, degreename, res);
 				sb.append(names+"\n");
 				score+=res;
-				
+
 			}
 			String average = String.format("The average score for this school is %.3f", score/count);
 			sb.append(average);
-			
+
 			if(count == 0){
 				return "No results for this school";
 			}
@@ -838,7 +881,7 @@ public class Queries {
 		}
 		return sb.toString();
 	}
-	
+
 	public String degreeResult(String s){
 		Double score = 0.0;
 		StringBuilder sb = new StringBuilder("");
@@ -852,23 +895,23 @@ public class Queries {
 			String query = "SELECT s.FIRSTNAME, s.LASTNAME, sd.RESULT FROM STUDENT_DEGREE AS sd INNER JOIN DEGREE AS d ON d.DEGREEID = sd.DEGREE INNER JOIN STUDENT AS s ON s.STUDENTID = sd.STUDENT WHERE d.DEGREENAME = '"+s+"' AND sd.RESULT IS NOT NULL";
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
-			
+
 			if(rs.next()){
 				count+=1;
 				String fname = rs.getString("FIRSTNAME");
 				String lname = rs.getString("LASTNAME");
 				String result = rs.getString("RESULT");
-				
+
 				String name = fname +" "+ lname;
 				Double res = Double.parseDouble(result);
 				String names = String.format(" %-50.50s %-10.3f\n", name, res);
 				sb.append(names+"\n");
 				score+=res;
-				
+
 			}
 			String average = String.format("The average score for this degree is %.3f", score/count);
 			sb.append(average);
-			
+
 			if(count == 0){
 				return "No results for that degree";
 			}
@@ -877,24 +920,32 @@ public class Queries {
 		}
 		return sb.toString();
 	}
-	
+
 	public String insertStudent(String name, String lname, String email, String address, String matric, String degree){
 		StringBuilder sb = new StringBuilder("");
+		int count = 0;
 		ResultSet rs;
 		try{
-			String query ="SELECT * FROM USER WHERE MATRICNO = '"+matric+"'";
+			String query ="SELECT s.FIRSTNAME, s.LASTNAME, us.MATRICNO FROM USER AS us INNER JOIN STUDENT AS s ON s.USERID = us.ID";
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
-			
-			if(rs.next()){
-				sb.append("Error");
-			}else{
+
+			while(rs.next()){
+				String fname = rs.getString("FIRSTNAME").toLowerCase();
+				String lstname = rs.getString("LASTNAME").toLowerCase();
+				String mat = rs.getString("MATRICNO").toLowerCase();
+				if((name.toLowerCase().equals(fname) && lstname.toLowerCase().equals(lstname)) || matric.toLowerCase().equals(mat)){
+					count+=1;
+					sb.append("Error");
+				}
+			}
+			if(count == 0){
 				String sql = "INSERT INTO USER VALUES ('"+countUsers()+"', '"+matric+"', '"+matric+"', 'Student')";
 				st.executeUpdate(sql);
-				
+
 				String sql1 = "INSERT INTO STUDENT VALUES ('"+countStudents()+"', '"+name+"', '"+lname+"', '"+email+"', '"+address+"', (SELECT ID FROM USER WHERE MATRICNO = '"+matric+"'))";
 				st.executeUpdate(sql1);
-				
+
 				String sql2 = "INSERT INTO STUDENT_DEGREE (STUDENT, DEGREE) VALUES ((SELECT s.STUDENTID FROM STUDENT AS s INNER JOIN USER AS us ON us.ID = s.USERID WHERE us.MATRICNO = '"+matric+"'), (SELECT DEGREEID FROM DEGREE WHERE DEGREENAME = '"+degree+"'))";
 				st.executeUpdate(sql2);
 			}
@@ -905,14 +956,14 @@ public class Queries {
 		}
 		return sb.toString();
 	}
-	
+
 	public int countUsers(){
 		ResultSet rs;
 		try{
 			String query = "SELECT COUNT (*) FROM USER";
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
-			
+
 			while(rs.next()){
 				return rs.getInt(1)+1;
 			}
@@ -923,14 +974,14 @@ public class Queries {
 		}
 		return 0;
 	}
-	
+
 	public int countStudents(){
 		ResultSet rs;
 		try{
 			String query = "SELECT COUNT (*) FROM STUDENT";
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
-			
+
 			while(rs.next()){
 				return rs.getInt(1)+1;
 			}
@@ -941,11 +992,24 @@ public class Queries {
 		}
 		return 0;
 	}
+
+	public String getScore(int id){
+		String s = "";
+		String[] score = {"A1","A2","A3","A4","A5","B1","B2","B3","C1","C2","C3","D1","D2","D3","E1","E2","E3","F1","F2","F3","G1","G2","H"};
+		int[] mark = {22,21,20,19,18,17,16,15,14,13,12,11,10,9,8,7,6,5,4,3,2,1,0};
+		for(int i = 0; i<score.length;i++){
+			if(id == mark[i]){
+				s+=score[i];
+			}
+		}
+
+		return s;
+	}
 	public String getUnique(){
 		StringBuilder sb = new StringBuilder("");
 		final String initial = "2";
 		sb.append(initial);
-		
+
 		Random r = new Random();
 		int val = r.nextInt(999999) + 100000;
 		sb.append(val);
@@ -959,9 +1023,9 @@ public class Queries {
 			String query = "SELECT us.MATRICNO FROM USER AS us INNER JOIN STUDENT AS s ON s.USERID = us.ID WHERE s.FIRSTNAME = '"+s+"' AND s.LASTNAME = '"+s1+"'";
 			st = conn.createStatement();
 			rs = st.executeQuery(query);
-			
+
 			if(rs.next()){
-				
+
 				m+=rs.getString("MATRICNO");
 			}
 			rs.close();
@@ -970,6 +1034,50 @@ public class Queries {
 			e.printStackTrace();
 		}
 		return m;
+	}
+
+	public String checkString(String s){
+		String s1 = "";
+		Pattern pattern = Pattern.compile("[a-zA-Z\\s-]*");
+		Matcher matcher = pattern.matcher(s);
+		boolean check = matcher.matches();
+		if(!check){
+			s1+="Exists";
+		}
+		return s1;
+	}
+
+	public String checkLastName(String s){
+		String s1 = "";
+		Pattern pattern = Pattern.compile("[a-zA-Z\\-]*");
+		Matcher matcher = pattern.matcher(s);
+		boolean check = matcher.matches();
+		if(!check){
+			s1+="Exists";
+		}
+		return s1;
+	}
+
+	public String checkAddress(String s){
+		String s1 = "";
+		Pattern pattern = Pattern.compile("[a-zA-Z0-9\\s.,]*");
+		Matcher matcher = pattern.matcher(s);
+		boolean check = matcher.matches();
+		if(!check){
+			s1+="Exists";
+		}
+		return s1;
+	}
+
+	public String checkEmail(String s){
+		String s1 = "";
+		Pattern pattern = Pattern.compile("[a-zA-Z0-9\\@._-]*");
+		Matcher matcher = pattern.matcher(s);
+		boolean check = matcher.matches();
+		if(!check){
+			s1+="Exists";
+		}
+		return s1;
 	}
 
 	public Users getUser(){
